[FE] refactor : PaginationBar 개선
- uri 방식 외에도 사용될 수 있도록 확장
- PaginationNavigator -> PaginationBar 네이밍 수정
