+++ frontend/src/common/components/PaginationBar/index.stories.tsx
@@ -1,0 +1,17 @@
+import PaginationBar from '.';
+
+import { ComponentStory } from '@storybook/react';
+
+export default {
+  title: 'common/components/PaginationBar',
+  component: PaginationBar,
+};
+
+const Template: ComponentStory<typeof PaginationBar> = (args) => <PaginationBar {...args} />;
+
+export const DefaultPaginationBar = Template.bind({});
+
+DefaultPaginationBar.args = {
+  itemCountInPage: 5,
+  totalItemCount: 63,
+};
